SetUp cart anection address, voucher, payment

+++ app/src/main/java/com/example/foodeli/MySqlSetUp/Schemas/Cart/Body/AddToCartBody.java
@@ -3,9 +3,9 @@ package com.example.foodeli.MySqlSetUp.Schemas.Cart.Body;
 public class AddToCartBody {
 
     private int uid, pid;
-    private float quantity;
+    private int quantity;
 
-    public AddToCartBody(int uid, int pid, float quantity) {
+    public AddToCartBody(int uid, int pid, int quantity) {
         this.uid = uid;
         this.pid = pid;
         this.quantity = quantity;
