+++ app/src/main/java/com/example/foodeli/MySqlSetUp/Schemas/ShopProduct/Body/UpdateProductDataBody.java
@@ -9,7 +9,7 @@ public class UpdateProductDataBody extends ProductPostBody{
             int uid,
             int sid,
             float price,
-            float stock,
+            int stock,
             String name,
             String unit,
             String shortDesc,
