+++ app/src/main/java/com/example/foodeli/MySqlSetUp/Schemas/UserVoucher/Body/CheckCanUseVoucher.java
@@ -1,10 +1,10 @@
 package com.example.foodeli.MySqlSetUp.Schemas.UserVoucher.Body;
 
 public class CheckCanUseVoucher {
-    private float total, shipping;
+    private double total, shipping;
     private int vid;
 
-    public CheckCanUseVoucher(float total, float shipping, int vid) {
+    public CheckCanUseVoucher(double total, double shipping, int vid) {
         this.total = total;
         this.shipping = shipping;
         this.vid = vid;
